+++ NEWS.md
@@ -1,4 +1,5 @@
 * Version 0.5.0 (2015-mm-dd)
+    - Connection config can be an object or a connectionString
     - Propagate db_max_row_size from global settings to pg default's maxRowSize
       Allows to set the max length in bytes for a row